Tests/LibWeb/Text/expected/ShadowDOM/slotted-does-not-match-nested-slot.txt

direct slot color: rgb(0, 0, 0)
direct leaf color: rgb(0, 128, 0)
outer slot color: rgb(0, 0, 0)
outer leaf color: rgb(255, 0, 0)
l1 slot color: rgb(0, 0, 0)
l2 slot color: rgb(0, 0, 0)
l1 leaf color: rgb(0, 0, 255)
named slot bg: rgba(0, 0, 0, 0)
named slot border: rgb(0, 0, 0)
named slot outline: rgb(0, 0, 0)
named slot color: rgb(0, 0, 0)
named slot caret: rgb(0, 0, 0)
named leaf bg: rgb(10, 10, 10)
named leaf border: rgb(20, 20, 20)
named leaf outline: rgb(30, 30, 30)
fb inner slot color: rgb(0, 0, 0)
fb fallback color: rgb(0, 0, 0)
multi slot color: rgb(0, 0, 0)
multi leaf[0] color: rgb(255, 0, 0)
multi leaf[1] color: rgb(255, 0, 0)
multi leaf[2] color: rgb(255, 0, 0)
